Rename ReservationUtils to TimeUtils
Later more time related utility functions will be adde here
so the new name makes more sense.
Refs #40.
